+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/schemas.py
@@ -50,5 +50,10 @@ class Schemas(Enum):
     """
     The URL of the OpenTelemetry schema version 1.30.0.
     """
+
+    V1_31_0 = "https://opentelemetry.io/schemas/1.31.0"
+    """
+    The URL of the OpenTelemetry schema version 1.31.0.
+    """
     # when generating new semantic conventions,
     # make sure to add new versions version here.